Set refresh to false for offline reindex
By default, refresh param is true in all write API requests, which
ensures the reindex operation to wait for the relevant ES primary and
replica shards to refresh after each write operation. That is the
behavior we want for consistency during online operations, but it's
unnecessary and very slow for the bulk offline reindex use case. So,
set refresh param to false for offline reindex.

Change-Id: Ic454d7cffc00c93fa44e15be49ac7ad9d99c245d