Remove dependency on //java/com/google/gerrit/proto
In I9926fdce6d //java/com/google/gerrit/proto was exposed in the plugin
API so that this lib module doesn't need to depend on it any more. This
removes 1MB of size from the final module artifact, as protobuf library
is now also omitted from the shaded module jar file.

Change-Id: I217feca153e3900cc04831689faca0ce7f0fa237