ElasticIndexModule: Accept auto flush value exposed by gerrit core
By default, refresh param is true in all write API requests, which
ensures the reindex operation waits for the relevant ES primary and
replica shards to refresh after each write operation. That is the
behavior we want for consistency during online operations, but it's
unnecessary and very slow for the bulk offline reindex use case. So,
set refresh param to false for offline reindex.

Change-Id: I719ab0d39fe58188f5361bbcf9b282b42cc2cfaf