Use Elasticsearch's 'Point In Time' API for searches
Elasticsearch's PIT[1] (Point In Time) API was introduced in version
7.10. Searches with a large 'size' values (translates to 'limit' in
Gerrit queries) are observed to perform poorly. PIT's search_after[2]
can be used to paginate results. This will help with the aforementioned
case to do an initial search with a small size (say 100) and then
gradually increase the size to meet the provided query limit.

PIT pagination also helps overcome the default response buffer limit
(which is 100MB) set in the Elasticsearch's low-level REST client.
While the response buffer limit can be increased, it is likely that
the data nodes and client will become overloaded. For example, queries
like [3] reach this default buffer limit on most Gerrit instances.

PIT is enabled by default and can be turned off via a config setting.
Also, it is not used unless ES version is 7.10+

[1] https://www.elastic.co/guide/en/elasticsearch/reference/7.10/point-in-time-api.html
[2] https://www.elastic.co/guide/en/elasticsearch/reference/7.10/paginate-search-results.html#search-after
[3] gerrit query status:merged --no-limit

Change-Id: I8baa6e0725aba851e3acb9ac3159531623fe95c3