Use official elasticsearch docker image for tests
This is a switch from elasticsearch-oss image to the full image.
ElasticSearch's point in time feature will be introduced in a
follow-up change (I8baa6e07) and is not available in the oss
image. Hence the need for switch. When running tests with the
full image we will likely run into OOM issues on the docker
hosts. To help mitigate that, run the tests sequentially.

Change-Id: I951ce5e630a528f87ff3946647d96fe136702043