Update 'exact' field mapping to avoid duplicate data
As an example, the mapping for the exact field 'status' in 'changes'
index is updated as below to avoid duplication.

Without change:
  "status": {
    "type": "keyword",
    "fields": {
      "key": {
        "type": "keyword"
      }
    }
  }

With change:
  "status": {
    "type": "keyword"
  }

On a Gerrit instance with ~4M changes, this change brings down the
size of changes index from ~25G to ~20G.

Release-Notes: Size of indices is reduced. Re-index is required.
Change-Id: I680eb01c7e8c96548ddb4b321f383916f165af28