Use 'index_prefixes' for prefix type fields
Performance of prefix searches[1] can be improved using
index_prefixes[2]. The min and max chars are configured
to 5 and 19 respectively. As this change updates the index
mapping, a re-index will be required to see a performance
improvement.

On a Gerrit site with ~4M changes and size of the changes
index is ~25G, a query like 'gerrit query Ica34c2c3bb'
which translates to prefix query[3] in Elasticsearch now
takes ~7ms with this change as opposed to ~9ms without it.

[1] https://www.elastic.co/guide/en/elasticsearch/reference/7.17/faster-prefix-queries.html
[2] https://www.elastic.co/guide/en/elasticsearch/reference/7.17/index-prefixes.html
[3] "query":{"match_phrase_prefix":{"change_id":"Ica34c2c3bb"}}

Release-Notes: performance of prefix queries is improved (after re-index)
Change-Id: I37d58919b6e8dc5caaa8950b45593042980d0bda