Deny usage of this elasticsearch module with paginationType set to NONE
The paginationType NONE option needs to be honoured
by the indexing backend [1].
The elasticsearch libModule should throw an IAE to highlight
that the backend doesn't support NONE pagination type.
In particular setting the index.max_result_window in
elasticsearch [2] cause the queries to fail if maxLimit [3]
is not set, or the results to be truncated if maxLimit is
set to the same value of the index.max_result_window.

[1]: https://gerrit-review.googlesource.com/c/gerrit/+/377794
[2]: https://www.elastic.co/guide/en/elasticsearch/reference/current/index-modules.html#dynamic-index-settings
[3]: https://gerrit-documentation.storage.googleapis.com/Documentation/3.6.6/config-gerrit.html#index.maxLimit

Bug: Issue 291106652
Release-Notes: skip
Change-Id: Ic72886976568f6865d5b9d1e6ffa2d4ef1283bac