Failfast when elasticsearch module has paginationType set to NONE
The paginationType NONE option needs to be honoured by the indexing
backend (see change 377794).
The elasticsearch backend doesn't support this configuration therefore
it throws ProvisisionException during the startup when such
configuration is detected. Here is the exception message:

  The 'index.paginationType = NONE' configuration is not supported
  by Elasticsearch

Note: this change should be reverted in master when the support of
'index.paginationType=NONE' through the 'point-in-time' API is added
(see [1] for details).

[1]: https://www.elastic.co/guide/en/elasticsearch/reference/7.10/point-in-time-api.html

Bug: Issue 291106652
Change-Id: I47c6b7f972e2b1d443cc5748386290dc10556f2a