test: Support containers with SSL
Most of the test code assumed the test container would always be
unsecured (reachable via 'http://', no username/password, etc), but it's
possible (and the default in newer Elasticsearch versions) to configure
the container to run in a secure mode.

This change ensures the tests use any passwords and self-signed
certificates generated by the test container, as well as correctly using
'https://' URLs when needed.

Change-Id: Ifca5948995ec881d8e941702fef611a3455e0cb7