test: Always enable SSL for ES containers
Most of the test code assumed the test container would always be
unsecured (reachable via 'http://', no username/password, etc), but it's
possible (and the default in newer Elasticsearch versions) to configure
the container to run in a secure mode. Use the new support in
testcontainers to always enable SSL for ES so that Elasticsearch
versions will be tested similarly.

This change ensures the tests use any passwords and self-signed
certificates generated by the test container, as well as correctly using
'https://' URLs when needed.

This change uses the tests in testcontainers [1] as an example for how
to do this and ensures the docker-in-docker setup used by gerrit-ci is
supported.

[1] https://github.com/testcontainers/testcontainers-java/blob/main/modules/elasticsearch/src/test/java/org/testcontainers/elasticsearch/ElasticsearchContainerTest.java

Change-Id: Ifca5948995ec881d8e941702fef611a3455e0cb7