Provide a saner default for http client thread count
The default for HttpAsyncClientBuilder is the number of available
processors and the ES RestClientBuilder sets up a pool of 30 maximum
connections. IO threads multiplex connections, so we don't need anything
close to 1:1. Use a safe default of 1 IO thread for every 2 connections,
but keep the number of processors as an upper bound.

Change-Id: I0e4b2f604560fd0414c52d296d6c6a967e6587c9